requirements.txt
streamlit
sentence-transformers
scipy
numpy
PyPDF2
python-docx
pandas
